Resume Page
Got a copy of my most revised version of my resume up on the site! I used a module called React-pdf which had some side effects with it that took some testing and reading of the docs. I figured out also how to make a download able link which was pretty cool!